Refactor(helm): extract release dispatcher into release.go
Move installOrUpgradeChart (dispatcher between install/upgrade
   with fingerprint dedup, adoption detection, and publishVersion pin
  short-circuit) and computeReleaseFingerprint out of helm.go.

No behavior change. All 240 unit specs remain green.
